store: Move 'Catalog' to its own module
